feat(unityengine): fix labels and some d3d11on12 games maybe

+++ src/games/unityengine/Plugin/Beautify/Kronnect_Beautify_ACES_0x8EFD32D4.ps_4_0.hlsl
@@ -148,7 +148,7 @@ void main(
                                                                      : renodx::tonemap::config::hue_correction_type::CUSTOM;
   config.hue_correction_strength = injectedData.toneMapHueCorrection;
   config.hue_correction_color = lerp(r0.xyz, hueCorrectionColor, injectedData.toneMapHueShift);
-  config.reno_drt_hue_correction_method = (int)injectedData.toneMapHueProcessor;
+  config.reno_drt_hue_correction_method = injectedData.toneMapHueProcessor;
   config.reno_drt_tone_map_method = injectedData.toneMapType == 4.f ? renodx::tonemap::renodrt::config::tone_map_method::REINHARD
                                                                     : renodx::tonemap::renodrt::config::tone_map_method::DANIELE;
   config.reno_drt_working_color_space = 0;
